Removed couple of usize and reduced memory footprint

+++ src/arena.c
@@ -4,7 +4,7 @@
 #define DEFAULT_ALIGNMENT  (2 * sizeof(uint16))
 #define is_power_of_two(x) ((x != 0) && ((x & (x - 1)) == 0))
 
-void arena_init(Arena *a, void *buffer, usize buffer_size)
+void arena_init(Arena *a, void *buffer, uint32 buffer_size)
 {
     a->buf         = buffer;
     a->size        = buffer_size;
@@ -12,7 +12,7 @@ void arena_init(Arena *a, void *buffer, usize buffer_size)
     a->curr_offset = 0;
 }
 
-static uintptr_t align_forward(uintptr_t ptr, usize align)
+static uintptr_t align_forward(uintptr_t ptr, uint32 align)
 {
     uintptr_t p      = ptr;
     uintptr_t a      = (uintptr_t)align;
@@ -25,7 +25,7 @@ static uintptr_t align_forward(uintptr_t ptr, usize align)
     return p;
 }
 
-static void *arena_alloc_align(Arena *a, usize size, usize align)
+static void *arena_alloc_align(Arena *a, uint32 size, uint32 align)
 {
     uintptr_t curr_ptr = (uintptr_t)a->buf + (uintptr_t)a->curr_offset;
     uintptr_t offset   = align_forward(curr_ptr, align);
@@ -54,7 +54,7 @@ uintptr_t arena_current_offset(const Arena *a)
     return offset;
 }
 
-void *arena_alloc(Arena *a, usize size) { return arena_alloc_align(a, size, DEFAULT_ALIGNMENT); }
+void *arena_alloc(Arena *a, uint32 size) { return arena_alloc_align(a, size, DEFAULT_ALIGNMENT); }
 void *arena_at(const Arena *a, uintptr_t offset) { return &a->buf[offset]; }
 
 void arena_reset(Arena *a)
